Added Error Dialog and Confirm Dialog, Fixed Double Join
Added both the general Error dialog for handling site-wide errors that will give the user information to correct the error themselves (such as logging in) and the Confirm dialog which keeps the user from performing important actions accidentally (such as leaving a party).

Also edited multiple files to disallow an account to join the same party multiple times.  Tidied up some other rough points in the code as well and added imagery for the character screen.