Search/Filter Implementation
Added functionality to partydirectory component allowing the user to send requests to backend to filter results based on the name of the instance or purpose of the party.  It became apparent that the myparties component was no longer necessary, as the user should be able to tick a checkbox allowing them to filter for parties they either own or have joined, so deleted myparties component and removed all associations from project.